+++ src/commons/s3.rs
@@ -164,10 +164,13 @@ pub struct S3ConnectionSpec {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub port: Option<u16>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub access_style: Option<S3AccessStyle>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub secret_class: Option<String>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub tls: Option<Tls>,
 }
+
 impl S3ConnectionSpec {
     /// Convenience function to retrieve the spec of a S3 connection resource from the K8S API service.
     pub async fn get(
@@ -197,9 +200,22 @@ impl S3ConnectionSpec {
     }
 }
 
+#[derive(strum::Display, Clone, Debug, Deserialize, JsonSchema, PartialEq, Serialize)]
+#[strum(serialize_all = "PascalCase")]
+pub enum S3AccessStyle {
+    Path,
+    VirtualHosted,
+}
+
+impl Default for S3AccessStyle {
+    fn default() -> Self {
+        S3AccessStyle::VirtualHosted
+    }
+}
+
 #[cfg(test)]
 mod test {
-    use crate::commons::s3::S3ConnectionDef;
+    use crate::commons::s3::{S3AccessStyle, S3ConnectionDef};
     use crate::commons::s3::{S3BucketSpec, S3ConnectionSpec};
 
     #[test]
@@ -210,6 +226,7 @@ mod test {
                 host: Some("host".to_owned()),
                 port: Some(8080),
                 secret_class: None,
+                access_style: Some(S3AccessStyle::VirtualHosted),
                 tls: None,
             })),
         };
@@ -222,6 +239,7 @@ connection:
   inline:
     host: host
     port: 8080
+    accessStyle: VirtualHosted
 "
             .to_owned()
         )